@@ -1,11 +1,8 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp, h } from 'vue'
 import store from './store'
+import App from './App.vue'
 
-Vue.config.productionTip = false
-
-const app = new Vue({
-  render: h => h(App),
-}).$mount('#app')
-
-app.use(store)
+const app = createApp({
+	render: () => h(App)
+})
+app.use(store).mount('#app')
